properly placed code and templates
