--- base.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='custom.css') }}">
-    <title>Your App</title>
-</head>
-<body>
-    <nav>
-        <!-- Your navbar content -->
-        <a href="{{ url_for('index') }}">Home</a>
-        <a href="{{ url_for('add_factor') }}">Add Factor</a>
-    </nav>
-    <div class="content">
-        {% block content %}{% endblock %}
-    </div>
-</body>
-</html>
