+++ flask_app/templates/base.html
@@ -13,6 +13,7 @@
 
     {%- block styles %}
     <link rel="icon" type="image/x-icon" href="/static/assets/favicon.ico">
+    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='custom.css') }}">
     <!-- Bootstrap -->
     <link href="{{bootstrap_find_resource('css/bootstrap.css', cdn='bootstrap')}}" rel="stylesheet">
     <link rel="stylesheet" href="/static/css/style.css?version={{ now }}">
@@ -33,6 +34,15 @@
     <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
     {%- endblock scripts %}
     {%- endblock body %}
+
+    <nav>
+      <!-- navbar content -->
+      <a href="{{ url_for('index') }}">Home</a>
+      <a href="{{ url_for('add_factor') }}">Add Factor</a>
+    </nav>
+    <div class="content">
+        {% block content %}{% endblock %}
+    </div>
     </body>
     {%- endblock html %}
