+++ flask_app/templates/navbar.html
